+++ OpenGL-basico/geometry/vector3.h
@@ -4,40 +4,40 @@
 
 struct vector3
 {
-    float x, y, z;
+    double x, y, z;
 
-    float length = x * x + y * y + z * z;
-    float magnitude = std::sqrt(length);
+    double length = x * x + y * y + z * z;
+    double magnitude = std::sqrt(length);
 
-    vector3(const float x = 0.0, const float y = 0.0, const float z = 0.0) : x(x), y(y), z(z)
+    vector3(const double x = 0.0, const double y = 0.0, const double z = 0.0) : x(x), y(y), z(z)
     {
     }
 
     // Getter methods for the coordinates
-    float get_x() const;
-    float get_y() const;
-    float get_z() const;
+    double get_x() const;
+    double get_y() const;
+    double get_z() const;
 
 
     // Setter methods for the coordinates
-    void set_x(float new_x);
-    void set_y(float new_y);
-    void set_z(float new_z);
+    void set_x(double new_x);
+    void set_y(double new_y);
+    void set_z(double new_z);
 
     // Operators
     vector3 operator+(const vector3& other) const;
     void operator+=(const vector3& other);
     vector3 operator -(const vector3& other) const;
     void operator-=(const vector3& other);
-    vector3 operator*(float s) const;
+    vector3 operator*(double s) const;
     vector3 operator*(const vector3& other) const;
     vector3 operator/(int s) const;
     vector3 operator-() const;
 
     bool operator==(const vector3& zero) const;
 
-    float dot_product(const vector3& other) const;
-    float angle_with(const vector3& other) const; // returns in radians
+    double dot_product(const vector3& other) const;
+    double angle_with(const vector3& other) const; // returns in radians
 
     vector3 normalize() const;
     vector3 symmetrical(const vector3& other) const;
@@ -46,5 +46,5 @@ struct vector3
     void print() const;
     void reset();
 
-    float get_norm() const;
+    double get_norm() const;
 };
