+++ OpenGL-basico/ray-tracing/sphere.h
@@ -4,13 +4,13 @@
 
 class sphere : public object
 {
-	float radius_;
+	double radius_;
 public:
-    sphere(vector3 pos, float radius, color color, float reflectivity, float shininess) : object(pos, color, reflectivity, shininess), radius_(radius)
+    sphere(vector3 pos, double radius, color color, double reflectivity, double shininess) : object(pos, color, reflectivity, shininess), radius_(radius)
     {
     }
 	
     bool test_intersection(ray &rayo, vector3 &point, vector3 &normal) override;
-	float get_radius() const;
+	double get_radius() const;
 	~sphere() override;
 };
